+++ include/modules.h
@@ -286,9 +286,10 @@ struct Cmode {
 	/** Free and remove parameter from list.
 	 * This function pointer is NULL (unused) for modes without parameters.
 	 * @param parastruct		The parameter struct
+	 * @returns Normally return 0, must return 1 if it 'resisted' freeing.
 	 * @note In most cases you will just call safe_free() on 'list'
 	 */
-	void (*free_param)(void *parastruct);
+	int (*free_param)(void *parastruct);
 
 	/** duplicate a struct and return a pointer to duplicate.
 	 * This function pointer is NULL (unused) for modes without parameters.
@@ -347,7 +348,7 @@ typedef struct {
 	void *		(*put_param)(void *, const char *);
 	const char *	(*get_param)(void *);
 	const char *	(*conv_param)(const char *, Client *, Channel *);
-	void		(*free_param)(void *);
+	int		(*free_param)(void *);
 	void *		(*dup_struct)(void *);
 	int		(*sjoin_check)(Channel *, void *, void *);
 	char		local;